requirements.txt
celery~=4.2
Django~=2.1
djangorestframework~=3.8
psycopg2-binary~=2.7
